+++ src/main/java/app/bankApp/serviceBank/CreditCalculationPayment.java
@@ -11,7 +11,7 @@ public class CreditCalculationPayment {
      * @param term for the entite period
      * @return double payment
      */
-    public double calc(int amount, double ptc, int term){
+    public double calc(double amount, double ptc, int term){
 
         double ptcMonth = (ptc/12)/100;//месячная процентная ставка
         double k = Math.pow((1 + ptcMonth), term); // возведение в степень
